Renamed items, moved stuff around, updated some names
